+++ rustfs/src/server/event.rs
@@ -18,13 +18,10 @@ use super::{
 };
 use crate::init::reconcile_persisted_bucket_notification_configurations;
 use crate::storage_api::server::event::{
-    EventArgs as EcstoreEventArgs, StorageObjectInfo, read_existing_server_config_no_lock, register_event_dispatch_hook,
+    EventArgs as EcstoreEventArgs, read_existing_server_config_no_lock, register_event_dispatch_hook,
     with_server_config_read_lock,
 };
-use jiff::Timestamp;
-use rustfs_notify::{
-    EventArgs as NotifyEventArgs, NotificationError, NotificationRuntimeState, NotificationSystem, NotifyObjectInfo,
-};
+use rustfs_notify::{EventArgs as NotifyEventArgs, NotificationError, NotificationRuntimeState, NotificationSystem};
 use rustfs_s3_types::EventName;
 use std::future::Future;
 use std::net::SocketAddr;
@@ -81,30 +78,7 @@ pub fn is_notify_module_enabled() -> bool {
     NOTIFY_MODULE_ENABLED.load(Ordering::Relaxed)
 }
 
-pub(crate) fn convert_ecstore_object_info(object: StorageObjectInfo) -> NotifyObjectInfo {
-    NotifyObjectInfo {
-        bucket: object.bucket,
-        name: object.name,
-        size: object.size,
-        etag: object.etag,
-        content_type: object.content_type,
-        user_defined: object
-            .user_defined
-            .iter()
-            .map(|(key, value)| (key.clone(), value.clone()))
-            .collect(),
-        version_id: object.version_id.map(|version_id| version_id.to_string()),
-        mod_time: object.mod_time.and_then(offset_date_time_to_timestamp),
-        restore_expires: object.restore_expires.and_then(offset_date_time_to_timestamp),
-        storage_class: object.storage_class,
-        transitioned_tier: (!object.transitioned_object.tier.is_empty()).then_some(object.transitioned_object.tier),
-    }
-}
-
-fn offset_date_time_to_timestamp(value: time::OffsetDateTime) -> Option<Timestamp> {
-    let nanosecond = value.nanosecond().try_into().ok()?;
-    Timestamp::new(value.unix_timestamp(), nanosecond).ok()
-}
+pub(crate) use crate::shared_types::convert_ecstore_object_info;
 
 fn convert_ecstore_event_args(args: EcstoreEventArgs) -> Option<NotifyEventArgs> {
     let version_id = args.object.version_id.map(|v| v.to_string()).unwrap_or_default();
