+++ scripts/layer-dependency-baseline.txt
@@ -15,22 +15,14 @@ cycle|composition<->interface
 cycle|infra<->interface
 dep|rustfs/src/admin/handlers/scanner.rs|interface->composition|crate::startup_background::ENV_SCANNER_ENABLED
 dep|rustfs/src/admin/handlers/scanner.rs|interface->composition|crate::startup_background::scanner_enabled_from_env
-dep|rustfs/src/app/admin_usecase.rs|app->interface|crate::server::DependencyReadiness
 dep|rustfs/src/app/admin_usecase.rs|app->interface|crate::server::collect_dependency_readiness_report
 dep|rustfs/src/app/bucket_usecase.rs|app->interface|crate::admin::handlers::site_replication::site_replication_bucket_meta_hook
 dep|rustfs/src/app/bucket_usecase.rs|app->interface|crate::admin::handlers::site_replication::site_replication_delete_bucket_hook
 dep|rustfs/src/app/bucket_usecase.rs|app->interface|crate::admin::handlers::site_replication::site_replication_make_bucket_hook
-dep|rustfs/src/app/bucket_usecase.rs|app->interface|crate::server::RemoteAddr
-dep|rustfs/src/app/object_usecase.rs|app->interface|crate::server::convert_ecstore_object_info
-dep|rustfs/src/cluster_snapshot.rs|infra->interface|crate::server::DependencyReadiness
-dep|rustfs/src/cluster_snapshot.rs|infra->interface|crate::server::DependencyReadinessReport
-dep|rustfs/src/cluster_snapshot.rs|infra->interface|crate::server::ReadinessDegradedReason
 dep|rustfs/src/cluster_snapshot.rs|infra->interface|crate::server::snapshot_dependency_readiness_report
 dep|rustfs/src/runtime_sources.rs|infra->app|crate::app::context
-dep|rustfs/src/storage/access.rs|infra->interface|crate::server::RemoteAddr
 dep|rustfs/src/storage/ecfs_extend.rs|infra->interface|crate::server::cors
 dep|rustfs/src/storage/ecfs_extend.rs|infra->interface|crate::storage::ecfs::ListObjectUnorderedQuery
-dep|rustfs/src/storage/helper.rs|infra->interface|crate::server::convert_ecstore_object_info
 dep|rustfs/src/storage/helper.rs|infra->interface|crate::server::is_audit_module_enabled
 dep|rustfs/src/storage/helper.rs|infra->interface|crate::server::is_notify_module_enabled
 dep|rustfs/src/storage/helper.rs|infra->interface|crate::server::refresh_audit_module_enabled
